[Refactor] Move shared connector sink helpers to connector common
