+++ be/src/connector/partition_chunk_writer.h
@@ -21,9 +21,9 @@
 #include "column/vectorized_fwd.h"
 #include "common/thread/threadpool.h"
 #include "compute_env/load_spill/load_chunk_spiller.h"
+#include "connector/common/connector_sink_commit.h"
+#include "connector/common/connector_sink_profile.h"
 #include "connector/utils.h"
-#include "connector_primitive/connector_sink_commit.h"
-#include "connector_primitive/connector_sink_profile.h"
 #include "formats/file_writer.h"
 #include "fs/fs_fwd.h"
