+++ be/src/connector/partitioned_connector_chunk_sink.h
@@ -22,10 +22,10 @@
 
 #include "column/vectorized_fwd.h"
 #include "common/status.h"
+#include "connector/common/connector_sink_profile.h"
 #include "connector/partition_chunk_writer.h"
 #include "connector/utils.h"
 #include "connector_primitive/connector_sink.h"
-#include "connector_primitive/connector_sink_profile.h"
 #include "runtime/runtime_fwd.h"
 
 namespace starrocks::formats {
