+++ be/test/connector_primitive/CMakeLists.txt
@@ -17,6 +17,9 @@ if (APPLE)
     list(APPEND BE_TEST_MACOS_LINK_STUBS ../base/macos_link_stubs.cpp)
 endif()
 
+set(CONNECTOR_PRIMITIVE_TEST_DEPENDENCIES ${STARROCKS_DEPENDENCIES})
+list(REMOVE_ITEM CONNECTOR_PRIMITIVE_TEST_DEPENDENCIES arrow_testing)
+
 set(CONNECTOR_PRIMITIVE_TEST_LINK_LIBS
     ConnectorPrimitive
     ExecPrimitive
@@ -29,7 +32,7 @@ set(CONNECTOR_PRIMITIVE_TEST_LINK_LIBS
     Base
     Gutil
     StarRocksGen
-    ${STARROCKS_DEPENDENCIES}
+    ${CONNECTOR_PRIMITIVE_TEST_DEPENDENCIES}
     ${STARROCKS_STATIC_LINK_LIBS}
     ${WL_LINK_DYNAMIC} ${STARROCKS_SYSTEM_LINK_LIBS} ${WRAP_LINKER_FLAGS}
     ${WL_START_GROUP}
